User signup and login completed - token generation also completed

+++ models/user.go
@@ -5,14 +5,14 @@ type User struct {
 	Email    *string `json:"email" validate:"email,required"`
 	Password *string `json:"password" validate:"required,min=4"`
 	Phone    *string `json:"phone"`
-	BookID   *int16  `json:"book_id"`
-	UserID   *int16  `json:"user_id"`
+	BookID   *int    `json:"book_id"`
+	UserID   *int    `json:"user_id"`
 }
 
 type Book struct {
-	BookID    *int16  `json:"book_id"`
+	BookID    *int    `json:"book_id"`
 	Author    *string `json:"author" validate:"required"`
 	Publisher *string `json:"publisher"`
 	Title     *string `json:"title" validate:"required"`
-	UserID    *int16  `json:"user_id" validate:"required"`
+	UserID    *int    `json:"user_id" validate:"required"`
 }
